Backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pandas==2.2.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
python-multipart==0.0.9
pydantic==2.6.1
